[#143] refactor: data provider 인터페이스 삭제
