+++ api/src/main/java/org/silsagusi/api/agent/application/AgentService.java
@@ -3,9 +3,9 @@ package org.silsagusi.api.agent.application;
 import org.silsagusi.api.agent.application.dto.AgentDto;
 import org.silsagusi.api.agent.application.dto.AgentUpdateRequest;
 import org.silsagusi.api.agent.application.dto.UsernameDto;
+import org.silsagusi.api.agent.infrastructure.AgentDataProvider;
+import org.silsagusi.api.message.infrastructure.dataProvider.MessageTemplateDataProvider;
 import org.silsagusi.core.domain.agent.Agent;
-import org.silsagusi.core.domain.agent.AgentDataProvider;
-import org.silsagusi.core.domain.message.dataProvider.MessageTemplateDataProvider;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
