+++ api/src/main/java/org/silsagusi/api/agent/infrastructure/AgentDataProvider.java
@@ -5,7 +5,6 @@ import org.silsagusi.api.auth.jwt.RefreshTokenStore;
 import org.silsagusi.core.customResponse.exception.CustomException;
 import org.silsagusi.core.customResponse.exception.ErrorCode;
 import org.silsagusi.core.domain.agent.Agent;
-import org.silsagusi.core.domain.agent.AgentDataProvider;
 import org.springframework.stereotype.Component;
 
 import io.jsonwebtoken.Claims;
@@ -13,18 +12,16 @@ import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class AgentDataProviderImpl implements AgentDataProvider {
+public class AgentDataProvider {
 
 	private final AgentRepository agentRepository;
 	private final JwtProvider jwtProvider;
 	private final RefreshTokenStore refreshTokenStore;
 
-	@Override
 	public void createAgent(Agent agent) {
 		agentRepository.save(agent);
 	}
 
-	@Override
 	public void validateExist(Agent agent) {
 		if (agentRepository.existsByUsername(agent.getUsername())) {
 			throw new CustomException(ErrorCode.CONFLICT_USERNAME);
@@ -39,24 +36,20 @@ public class AgentDataProviderImpl implements AgentDataProvider {
 		}
 	}
 
-	@Override
 	public Agent getAgentById(Long agentId) {
 		return agentRepository.findById(agentId)
 			.orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
 	}
 
-	@Override
 	public Agent getAgentByNameAndPhone(String name, String phone) {
 		return agentRepository.findByNameAndPhone(name, phone)
 			.orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
 	}
 
-	@Override
 	public void updateAgent(Agent agent) {
 		agentRepository.save(agent);
 	}
 
-	@Override
 	public void deleteRefreshTokenByAccessToken(String accessToken) {
 		if (Boolean.FALSE.equals(jwtProvider.validateToken(accessToken))) {
 			return;
