+++ api/src/main/java/org/silsagusi/api/contract/infrastructure/ContractDataProvider.java
@@ -1,38 +1,36 @@
 package org.silsagusi.api.contract.infrastructure;
 
+import java.io.IOException;
+import java.time.LocalDate;
+import java.util.UUID;
+
+import org.silsagusi.core.customResponse.exception.CustomException;
+import org.silsagusi.core.customResponse.exception.ErrorCode;
+import org.silsagusi.core.domain.contract.entity.Contract;
+import org.silsagusi.core.domain.contract.info.ContractDetailInfo;
+import org.silsagusi.core.domain.contract.info.ContractInfo;
+import org.silsagusi.core.domain.contract.info.ContractSummaryInfo;
+import org.silsagusi.core.domain.customer.entity.Customer;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Component;
+import org.springframework.web.multipart.MultipartFile;
+
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 
 import lombok.RequiredArgsConstructor;
 
-import org.silsagusi.core.domain.contract.dataProvider.ContractDataProvider;
-import org.silsagusi.core.domain.contract.entity.Contract;
-import org.silsagusi.core.domain.contract.info.ContractDetailInfo;
-import org.silsagusi.core.domain.contract.info.ContractInfo;
-import org.silsagusi.core.domain.contract.info.ContractSummaryInfo;
-import org.silsagusi.core.domain.customer.entity.Customer;
-import org.silsagusi.core.customResponse.exception.CustomException;
-import org.silsagusi.core.customResponse.exception.ErrorCode;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Component;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-import java.time.LocalDate;
-import java.util.UUID;
-
 @Component
 @RequiredArgsConstructor
-public class ContractDataProviderImpl implements ContractDataProvider {
+public class ContractDataProvider {
 
 	private final ContractRepository contractRepository;
 
 	private final AmazonS3 amazonS3;
 	private static final String S3_BUCKET_NAME = "joonggaemoa";
 
-	@Override
 	public void createContract(Customer customerLandlord, Customer customerTenant, LocalDate createdAt,
 		LocalDate expiredAt, MultipartFile file) throws IOException {
 		String filename = fileUpload(file);
@@ -47,7 +45,6 @@ public class ContractDataProviderImpl implements ContractDataProvider {
 		contractRepository.save(contract);
 	}
 
-	@Override
 	public Page<ContractInfo> getAllContracts(Long agentId, Pageable pageable) {
 
 		Page<Contract> contractPage = contractRepository.findAllByCustomerLandlord_AgentId(agentId, pageable);
@@ -62,26 +59,22 @@ public class ContractDataProviderImpl implements ContractDataProvider {
 		return contractInfoPage;
 	}
 
-	@Override
 	public Contract getContract(String contractId) {
 		Contract contract = contractRepository.findById(contractId)
 			.orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_ELEMENT));
 		return contract;
 	}
 
-	@Override
 	public void validateAgentAccess(Long agentId, Contract contract) {
 		if (!contract.getCustomerLandlord().getAgent().getId().equals(agentId)) {
 			throw new CustomException(ErrorCode.FORBIDDEN);
 		}
 	}
 
-	@Override
 	public void deleteContract(Contract contract) {
 		contractRepository.delete(contract);
 	}
 
-	@Override
 	public ContractSummaryInfo getSummary(Long agentId) {
 
 		LocalDate today = LocalDate.now();
@@ -101,7 +94,6 @@ public class ContractDataProviderImpl implements ContractDataProvider {
 
 	}
 
-	@Override
 	public ContractDetailInfo getContractInfo(Contract contract) throws IOException {
 		ContractDetailInfo contractDetailInfo = ContractDetailInfo.of(contract);
 		contractDetailInfo.builder().url(getUrl(contractDetailInfo.getUrl()));
