+++ api/src/main/java/org/silsagusi/api/message/application/MessageService.java
@@ -3,13 +3,13 @@ package org.silsagusi.api.message.application;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import org.silsagusi.core.domain.agent.Agent;
-import org.silsagusi.core.domain.agent.AgentDataProvider;
-import org.silsagusi.core.domain.customer.dataProvider.CustomerDataProvider;
-import org.silsagusi.core.domain.customer.entity.Customer;
+import org.silsagusi.api.agent.infrastructure.AgentDataProvider;
+import org.silsagusi.api.customer.infrastructure.CustomerDataProvider;
 import org.silsagusi.api.message.application.dto.MessageDto;
 import org.silsagusi.api.message.application.dto.MessageUpdateRequest;
-import org.silsagusi.core.domain.message.dataProvider.MessageDataProvider;
+import org.silsagusi.api.message.infrastructure.dataProvider.MessageDataProvider;
+import org.silsagusi.core.domain.agent.Agent;
+import org.silsagusi.core.domain.customer.entity.Customer;
 import org.silsagusi.core.domain.message.entity.Message;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
