+++ api/src/main/java/org/silsagusi/api/message/infrastructure/repository/MessageTemplateRepository.java
@@ -1,4 +1,4 @@
-package org.silsagusi.api.message.infrastructure;
+package org.silsagusi.api.message.infrastructure.repository;
 
 import java.util.List;
