+++ api/src/main/java/org/silsagusi/api/notify/infrastructure/NotificationDataProvider.java
@@ -5,7 +5,6 @@ import java.util.List;
 
 import org.silsagusi.core.customResponse.exception.CustomException;
 import org.silsagusi.core.customResponse.exception.ErrorCode;
-import org.silsagusi.core.domain.notification.dataProvider.NotificationDataProvider;
 import org.silsagusi.core.domain.notification.entity.Notification;
 import org.silsagusi.core.domain.notification.entity.NotificationType;
 import org.springframework.stereotype.Component;
@@ -17,13 +16,12 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class NotificationDataProviderImpl implements NotificationDataProvider {
+public class NotificationDataProvider {
 
 	private static final Long DEFAULT_TIMEOUT = 60L * 1000 * 60;
 	private final EmitterRepository emitterRepository;
 	private final NotificationRepository notificationRepository;
 
-	@Override
 	public SseEmitter subscribe(Long agentId) {
 		SseEmitter emitter = new SseEmitter(DEFAULT_TIMEOUT);
 		emitterRepository.save(agentId, emitter);
@@ -36,7 +34,6 @@ public class NotificationDataProviderImpl implements NotificationDataProvider {
 		return emitter;
 	}
 
-	@Override
 	public void notify(Long agentId, NotificationType type, String content) {
 
 		boolean alreadyNotified = notificationRepository.existsByAgentIdAndTypeAndContent(agentId, type, content);
@@ -66,19 +63,16 @@ public class NotificationDataProviderImpl implements NotificationDataProvider {
 		}
 	}
 
-	@Override
 	public List<Notification> getAllByAgent(Long agentId) {
 		return notificationRepository.findAllByAgentId(agentId);
 	}
 
-	@Override
 	public Notification getNotification(Long notificationId) {
 		Notification notification = notificationRepository.findById(notificationId)
 			.orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_ELEMENT));
 		return notification;
 	}
 
-	@Override
 	public void markNotificationRead(Notification notification) {
 		notification.markRead();
 		notificationRepository.save(notification);
