+++ api/src/main/java/org/silsagusi/api/survey/application/SurveyService.java
@@ -3,19 +3,19 @@ package org.silsagusi.api.survey.application;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.silsagusi.api.agent.infrastructure.AgentDataProvider;
+import org.silsagusi.api.consultation.infrastructure.ConsultationDataProvider;
+import org.silsagusi.api.customer.infrastructure.CustomerDataProvider;
+import org.silsagusi.api.notify.infrastructure.NotificationDataProvider;
 import org.silsagusi.api.survey.application.dto.AnswerDto;
 import org.silsagusi.api.survey.application.dto.QuestionDto;
 import org.silsagusi.api.survey.application.dto.SurveyDto;
+import org.silsagusi.api.survey.infrastructure.dataProvider.SurveyDataProvider;
 import org.silsagusi.core.domain.agent.Agent;
-import org.silsagusi.core.domain.agent.AgentDataProvider;
-import org.silsagusi.core.domain.consultation.dataProvider.ConsultationDataProvider;
 import org.silsagusi.core.domain.consultation.entity.Consultation;
-import org.silsagusi.core.domain.customer.dataProvider.CustomerDataProvider;
 import org.silsagusi.core.domain.customer.entity.Customer;
-import org.silsagusi.core.domain.notification.dataProvider.NotificationDataProvider;
 import org.silsagusi.core.domain.notification.entity.NotificationType;
 import org.silsagusi.core.domain.survey.command.QuestionCommand;
-import org.silsagusi.core.domain.survey.dataProvider.SurveyDataProvider;
 import org.silsagusi.core.domain.survey.entity.Answer;
 import org.silsagusi.core.domain.survey.entity.Question;
 import org.silsagusi.core.domain.survey.entity.QuestionAnswerPair;
